test(126181): Correção dos testes após refatoração do código
